+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix typings which are broken in #31
   - The state now has the aborted flag.
   - This is technically a breaking change, but releasing it as a minor update.
+- Fix typings around Args
 
 ## [3.7.0] - 2019-11-09
 ### Changed